refactor: split main shrs crate into two crates
