+++ shrs_core/src/builtin/help.rs
