+++ shrs_core/src/builtin/history.rs
