+++ shrs_core/src/hooks.rs
