+++ shrs_core/src/prompt.rs
