+++ shrs_core/src/signal.rs
