+++ shrs_core/src/state.rs
